requirements.txt
Django==2.2.9
djangorestframework==3.10.2
django-choices==1.6.0
django-environ==0.4.3
kafka-python==1.4.7
python-logstash==0.4.6
uhashring==1.1
